# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aura Render API",
    "version": "1.0.0",
    "description": "\n# Aura Render API Documentation\n\nAura Render is an AI-powered video generation platform that transforms text into professional videos.\n\n## Features\n- **AI Content Analysis**: Intelligent analysis of input text for optimal video generation\n- **Material Matching**: Automatic matching and acquisition of relevant video materials\n- **Real-time Processing**: WebSocket-based real-time progress tracking\n- **Multi-format Support**: Support for various video formats and quality levels\n- **Enterprise Security**: JWT authentication, rate limiting, and audit logging\n- **Batch Processing**: Handle multiple video generation requests efficiently\n\n## Authentication\nMost endpoints require authentication using JWT tokens. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Rate Limiting\nAPI requests are subject to rate limiting:\n- **Free Users**: 10 requests/minute, 600 requests/hour\n- **Premium Users**: 100 requests/minute, 6000 requests/hour\n- **Enterprise**: Custom limits available\n\n## Error Handling\nAll errors follow a consistent format:\n```json\n{\n  \"error\": {\n    \"code\": \"ERROR_CODE\",\n    \"message\": \"Human-readable error message\",\n    \"details\": {\n      \"field\": \"Additional error context\"\n    }\n  }\n}\n```\n        ",
    "termsOfService": "https://aurarender.com/terms",
...
    },
...
    }
...
  "servers": [
...
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      }
    },
    "schemas": {
...
      "MaterialSearchRequest": {
        "type": "object",
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "example": "technology startup office teamwork",
            "description": "Search query for materials"
          },
          "type": {
            "type": "string",
            "enum": [
              "video",
              "image",
              "audio"
            ],
            "example": "video"
          },
          "duration_min": {
            "type": "number",
            "example": 5.0
          },
          "duration_max": {
            "type": "number",
            "example": 30.0
          },
          "resolution": {
            "type": "string",
            "enum": [
              "HD",
              "FHD",
              "4K"
            ],
            "example": "FHD"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "example": 20
          }
        }
      },
      "Material": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "mat_123456789"
          },
          "title": {
            "type": "string",
            "example": "Modern Office Teamwork"
          },
          "description": {
            "type": "string",
            "example": "Professional team collaborating in modern office space"
          },
          "type": {
            "type": "string",
            "enum": [
              "video",
              "image",
              "audio"
            ],
            "example": "video"
          },
          "url": {
            "type": "string",
            "example": "https://cdn.aurarender.com/materials/video_123.mp4"
          },
          "thumbnail_url": {
            "type": "string",
            "example": "https://cdn.aurarender.com/materials/thumb_123.jpg"
          },
          "duration": {
            "type": "number",
            "example": 15.3
          },
          "resolution": {
            "type": "string",
            "example": "1920x1080"
          },
          "file_size": {
            "type": "integer",
            "example": 5242880
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "office",
              "teamwork",
              "technology",
              "professional"
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/materials/search": {
      "post": {
        "tags": [
          "Materials"
        ],
        "summary": "Search for materials",
        "description": "Search for video, image, or audio materials",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "materials": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Material"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 156
                    },
                    "query_time": {
                      "type": "number",
                      "example": 0.045
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}